% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgeom_mle.R
\name{ci_bgeom}
\alias{ci_bgeom}
\title{Confidence intervals and prediction for beta-geometric glm's}
\usage{
ci_bgeom(mod, newdata, level = 0.05, type = c("prob", "response"))
}
\arguments{
\item{mod}{fitted `geomGLM` object}

\item{newdata}{optional `data.frame` of new predictor variable values}

\item{level}{confidence level}

\item{type}{whether to predict response or probability}
}
\value{
column vector of predictions
}
\description{
Calculate confidence intervals and prediction
}
