% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_hist_plot.R
\name{logistic_hist_plot}
\alias{logistic_hist_plot}
\title{Plot logistic data as paired histograms}
\usage{
logistic_hist_plot(data, x, xlab, xlog, ymin, ymax, formula)
}
\arguments{
\item{data}{simulatin output `data.frame`; must have a `speciation` column}

\item{x}{name (unquoted) of predictor variable}

\item{xlab}{label for x axis}

\item{xlog}{logical, should x-axis be log-transformed (default FALSE)}

\item{ymin}{optional minimum for y-axis}

\item{ymax}{optional maximum for y-axis}

\item{formula}{optional formula to pass to `geom_smooth`}
}
\description{
Plot logistic data as paired histograms
}
