% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgeom_mle.R
\name{lrt_bgeom}
\alias{lrt_bgeom}
\title{Likelihood ratio test for beta-geometric glm}
\usage{
lrt_bgeom(mod0, mod1)
}
\arguments{
\item{mod0}{reduced model}

\item{mod1}{full model}
}
\value{
an ANOVA table like object inheriting from `anova` and `data.frame`.
    The LR test statistic can be extracted with `$Chisq[2]` and the P-value
    extracted with \code{$`Pr(>Chisq)`}.
}
\description{
Preform likelihood ratio test on two models fitted with `glm_bgeom`
    distribution and logit link
}
