% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_binom.R
\name{pred_binom}
\alias{pred_binom}
\title{Custom predictions from a binomial glmmTMB model object}
\usage{
pred_binom(object, length_out = 50)
}
\arguments{
\item{object}{A fitted `glmmTMB` model object of binomial family.}

\item{length_out}{The length of the evenly spaced sequence
generated for each non-fixed predictor. Defaults to 50.}
}
\value{
A data frame combining the prediction grid (`newdata`) with
        the predicted probability (`fit`) and its standard error
        (`se_fit`)
}
\description{
Makes prediction on response scale across an evenly spaced sequence of
fixed effect in `object`, then generates response prediction and SE
without random effects (i.e. population-level sensu `?glmmTMB`).
}
