% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_sim.R
\name{r_sim}
\alias{r_sim}
\title{r_sim}
\usage{
r_sim(la, mu, g, m_prop, nu, tau, xi, np, nstep)
}
\arguments{
\item{la}{vector of bird rates}

\item{mu}{vector of death rates}

\item{g}{vector of immigration rates from global population}

\item{m_prop}{vector of proportions (local immigration):(global immigartion)}

\item{nu}{vector of rates of incipient speciation}

\item{tau}{waiting time to full speciation}

\item{xi}{vector of weights given to new immigrants in slowing progress to
full speciation}

\item{np}{number of local populations in the meta population}

\item{nstep}{scalar, number of simulation steps}
}
\value{
a matrix
}
\description{
A pure R implementation to simulate speciation in a meta-population
birth-death-immigration model; mostly for testing against rcpp version
}
