% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgeom_mle.R
\name{glm_bgeom}
\alias{glm_bgeom}
\title{Maximum likelihood estimation for beta geometric glm}
\usage{
glm_bgeom(formula, data, ...)
}
\arguments{
\item{formula}{an object of class "formula" that describes the model}

\item{data}{a data.frame where the data live (unlike `glm` and `lm`,
the data object is *required*)}

\item{...}{arguments passed to `control` of `optim`}
}
\value{
an object of class `glmBGeom` inspired by, but not strictly
    inheriting from, `glm` and `lm`
}
\description{
Find maximum likelihood parameters for glm assuming geometric
    distribution and logit link
}
\details{
Some methods that work for objects of class `glm` and `lm` will
    also work for the output of this function, specifically:
    `print`, `coef`, `logLik`, and `AIC`. A custom likelihood ratio test
    function is provided by `lrt_geom`
}
