% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_BDI_spec.R
\name{sim_BDI_spec}
\alias{sim_BDI_spec}
\title{Simulate a birth-death-speciation model}
\usage{
sim_BDI_spec(la, mu, g, m_prop, nu, tau, xi, np, nstep)
}
\arguments{
\item{la}{vector of bird rates}

\item{mu}{vector of death rates}

\item{g}{vector of immigration rates from global population}

\item{m_prop}{vector of proportions (local immigration):(global immigration)}

\item{nu}{vector of rates of incipient speciation}

\item{tau}{waiting time to full speciation}

\item{xi}{vector of weights given to new immigrants in slowing progress to
full speciation}

\item{np}{number of local populations in the meta population}

\item{nstep}{scalar, number of simulation steps}
}
\value{
a data.frame with columns for input parameter values and output
results
\itemize{
\item \code{la}
\item \code{mu}
\item \code{g}
\item \code{m_prop}
\item \code{nu}
\item \code{tau}
\item \code{xi}
\item \code{np}
\item \code{nstep}
\item \code{time}: simulation time step, if speciation occurred this number will
be \verb{< nstep}, if not this number will be \verb{== nstep}
\item \code{mean_pop_size}: the mean population size across all time steps and
populations
\item \code{speciation}: 0 (no speciation) or 1 (yes speciation)
}
}
\description{
A fast simulation of speciation in a meta-population
birth-death-immigration model
}
