% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_plot.R
\name{spec_plot}
\alias{spec_plot}
\title{Custom plotting routine for speciation as response variable}
\usage{
spec_plot(dat, x, xlab, xlog = FALSE, formula = NULL, ...)
}
\arguments{
\item{dat}{output from simulation}

\item{x}{name (unquoted) of predictor variable}

\item{xlab}{label for x axis}

\item{xlog}{logical, should x-axis be log-transformed (default FALSE)}

\item{formula}{optional formula to pass to \code{geom_smooth}}

\item{...}{additional arguments to pass to \code{geom_smooth}}
}
\value{
a \code{ggplot} object
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Plot speciation against a predictor variable with glm prediction
}\if{html}{\out{</div>}}
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Not a general-purpose function, use only for intended scenarios
}\if{html}{\out{</div>}}
}
