% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_plot.R
\name{spec_plot}
\alias{spec_plot}
\title{Custom plotting routine for speciation as response variable}
\usage{
spec_plot(dat, x, xlab, xlog = FALSE, formula = NULL, ...)
}
\arguments{
\item{dat}{output from simulation}

\item{x}{name (unquoted) of predictor variable}

\item{xlab}{label for x axis}

\item{xlog}{logical, should x-axis be log-transformed (default FALSE)}

\item{formula}{optional formula to pass to `geom_smooth`}

\item{...}{additional arguments to pass to `geom_smooth`}
}
\value{
a `ggplot` object
}
\description{
Plot speciation against a predictor variable with glm prediction
}
\details{
Not a general-purpose function, use only for intended scenarios
}
